Fix #1611 Rare crash when cycling items

+++ src/main/java/mezz/jei/gui/ingredients/CycleTimer.java
@@ -23,8 +23,8 @@ public class CycleTimer {
 		if (list.isEmpty()) {
 			return null;
 		}
-		Long index = ((drawTime - startTime) / cycleTime) % list.size();
-		return list.get(index.intValue());
+		long index = ((drawTime - startTime) / cycleTime) % list.size();
+		return list.get(Math.toIntExact(index));
 	}
 
 	public void onDraw() {
